Start implementing paths
This purpose of this branch is to refactor the current shape primitives
to be composed of paths, with paths becoming the new primitives.

A path is a sequence of path elements. A path element is a line or a
bezier curve.

This refactoring also introduces envelopes (copied from Haskell's
diagrams package) which generalise bounding boxes. This allows for more
flexible layout.

Once this refactoring is done we can add padding and 'at' layout, and
this will complete a usuable version of Doodle (though rotation remains
another attractive addition).